+++ src/models/__init__.py
@@ -1,5 +1,5 @@
 from .transformer_core import TransformerCore
-from .transformer_model import Transformer
-from .ft_nat import FTNAT
-from .refine_nat import RefineNAT
-from .cmlm_nat import CMLM
+from .transformer import Transformer
+from .ftnat import FTNAT
+from .refinenat import RefineNAT
+from .cmlm import CMLM
