+++ src/modules/__init__.py
@@ -1,4 +1,4 @@
 from .connections import ResidualConnection, HighwayConnection
-from .layersNAT import DecoderLayerNAT, DecoderNAT
+from .layers_nat import DecoderLayerNAT, DecoderNAT
 from .pooling import Fertility, Pooler
-from .PositionalEncoding import PositionalEncoding
+from .positional_encoding import PositionalEncoding
